drivers: i2s: i2s_nrfx: support MCK frequency constraints
When MCK is routed to an external device, we need to ensure that the MCK
frequency is acceptable to the device. This patch introduces the
mck-min-frequency and mck-max-frequency device tree properties, which
constrain the driver's clock configuration.

As an example, the SGTL5000 requires an MCK frequency between 8 and 27
MHz. On the nRF53, with a 11.2896 MHz ACLK source and a 44.1 kHz
sampling frequency, the driver will currently configure MCK at 1.4112
MHz. With this patch and the appropriate device tree configuration, the
driver will instead choose to pass through ACLK to MCK with no divider,
satisfying the constraints.

Signed-off-by: Ben Wolsieffer <benwolsieffer@gmail.com>